Module folding refactor.
