+++ src/app/patry-shop/components/patry-shop/patry-shop.component.spec.ts
